LPS-196671 Make class ImageProcessorImpl an internal class of class AMThumbnailsOSGiCommandsTest because it is only used here and and take commit  5960d98d5379522c01302a6db0c30c9e078830e3 into account.
